jamwork

+++ DataWH/WebApplication1/WebApplication1/Datamodel/RoomMeasurement.cs
@@ -5,9 +5,9 @@ namespace WebApplication1.Datamodel
     public class RoomMeasurement
     {
         [Key] public int Id { get; set; }
-        public int Light { get; set; }
-        public int Temperature { get; set; }
-        public int Humidity { get; set; }
-        public int Co2 { get; set; }
+        public decimal Light { get; set; }
+        public decimal Temperature { get; set; }
+        public decimal Humidity { get; set; }
+        public decimal Co2 { get; set; }
     }
 }
